5948 - full code review & fixes
